commit cloud: implement get_smartlog_by_version
Summary: This retrieves a workspace history filtering by either timestamp or version, and once it has the older version it forms a smartlog view with the contents. For this latest step we need to cast local bookmarks and remote bookmarks into maps so when forming a smartlog view we can determine wether a commit is part of a bookmark or not.

Reviewed By: muirdm

Differential Revision: D61443756

fbshipit-source-id: 7d0eab09d00d1df4f46f2ffcba728509c8d780a1