scs_server: use `SyncedCommitMapping::get_maybe_stale` in `commit_lookup_pushrebase_history` method
Summary:
When calling the `SyncedCommitMapping::get` method, we first query a DB replica, and if the DB replica returns no mappings we go to the DB master to make sure that the lack of mappings is not due to DB replication lag. This means that for commits that actually have no mappings we always query the DB master which increases the chances of overloading it.

For the `commit_lookup_pushrebase_history` SCS method, the replication lag that's usually < 0.5 seconds shouldn't affect the output much, and this method is meant to be best effort anyways so it's not worth overloading our DB just to ensure it's correctness. Let's use `SyncedCommitMapping::get_maybe_stale` and avoiding querying the DB master.

Reviewed By: singhsrb

Differential Revision: D61600540

fbshipit-source-id: 113f65d64cad8567c4595852a1c2b2657e34d909