mercurial/mutation: support missing primordials in fetch and store
Summary:
We are deprecating the use of primoridal commits.  Currently they are not used, but they are still maintained.

In this diff, we add support for handling if they are missing.

By default the mutation store will search for missing primordials if it cannot immediately determine the primordial by fetching the predecessor.  To prevent this search from triggering every time we encounter a missing primordial, we will roll out in the following steps:

* Add support for missing primordials, but continue to search for missing primordials, and continue to write them. **(This diff)**
* Once this code is fully released, flip the `scm/mononoke:mutation_find_missing_primordials` JK.  We will continue to write them when they are available, but won't search for them if they are missing.
* Disable writing them completely. (Next diff)

Reviewed By: andreacampi

Differential Revision: D61133826

fbshipit-source-id: 88c723178a0751dcea9c09d4123dd0a1bf3ce525