Fix submodule bug while deriving GitDeltaManifestV2 in gitimport
Summary:
There is an edge case when importing a Git repo where we try to derive `GitDeltaManifest` for a commit that changes an existing submodule into a regular file. In this type of change, the previous value for the entry in `GitTree` manifest is of type `submodule` but the new value has a different type.

Before deciding if we need to include an entry or not, we look at the new value of the entry and if that's a submodule we skip it. However, we perform no checks on the old value of the entry. The derivation fails when it tries to generate a delta entry by fetching both the old and new value and realizes that no object (blob or tree) exists for the old value.

This diff fixes that problem by checking the type of the old entry as well as the new entry.

Differential Revision: D61665202

fbshipit-source-id: 89ff67abd277c2dcd011a9d656732a6e411f8f2b