respond with exit message synchronously
Summary:
In the previous diff, we dealt with an out of order exit message. One possible albeit unlikely cause could be from `execution.on('exit')` differing from the exit criteria for `await execution`. If so, sending `onProgress('exit')` synchronously after await execution means we're guaranteed to send the onProgress `exit` message before dequeuing the next operation.

It's a little silly that we have 3 places doing execa handling these messages in very similar ways. We should probably extract this into something common so we're definitely consistent about this kind of thing.

Reviewed By: muirdm

Differential Revision: D61635300

fbshipit-source-id: 67513d3a9ab7ee33f3523774eeaa9747df8925d0