Fix out of order exit messages permanently breaking operation history
Summary:
Tricky bug: some users were seeing the "You are here" commit getting stuck somewhere, and not updating properly from that point on. `sl` on the CLI would show the current isDot commit, but the UI showed some old commit.

By looking through a bug report UI state, I noticed an operation in the history which did not have hasCompletedOptimisticState nor exitCode, despite being in the middle of the operation history stack. I think the current model intended that only the `currentOperation` should possibly not have exitCode filled out.

https://www.internalfb.com/phabricator/paste/view/P1538009735?lines=4615

I found that the operation state that sets the exit code only tries to set it on the currentOperation. But, a spawn event could move an operation to the operationHistory without setting the exit data.

Thus, it must be that the next operation's `spawn` message comes in BEFORE the previous event's `exit` notification. I didn't trace down exactly when or why this would happen, but I assume it's something about the way processes are reporting their `exit` events through execa.

Anyway, we can work around this by having the exit event on the client also lookup historical commands and not just the current operation.

Although I never got a repro of this issue, I believe this is a very plausible cause of the bug.

tbh, I don't like the whole "currentOperation" vs "operationHistory" thing. I don't remember why we did that. It would likely be simpler to keep it all in one list and know that the last entry is the currentOperation which may have not exited yet. We could have a selector if we need to rerender based only on the topmost entry.

Reviewed By: muirdm

Differential Revision: D61572500

fbshipit-source-id: 028f6d7345bcf817a9ecbf5e33579e259add6016