Integration test demonstrating failure while importing repo with invalid HEAD symref
Summary:
If a repo has `HEAD` symref pointing to a branch that does not exist for that repo, Metagit simply ignores it and proceeds as if the `HEAD` symref doesn't exist at all. Mononoke does the sane thing and error out but unfortunately that makes us diverge from the current behavior. This is why I plan to fix this to maintain parity with Metagit.

This diff has the integration test that demonstrates this problem.

Differential Revision: D61666112

fbshipit-source-id: b5ef4d2db085239b57560f88ad9cac45c644ab29