assert_status should return the number of tries
Summary:
# Background

Since we added the status cache which checks the Journal sequence to decide if a cached status can be reused instead of computing the diff every time `hg status` is called, there is a chance that the Journal sequence fails to advance in time so the old status is returned. This is the cost we have to pay.

To make the status tests more stable, we added the timeout parameter for `assert_status` helper function and made it by default 1 sec.

For the new tests added for status cache, we are relying on checking the counters to make sure the cache is working as expected. However, the retries from this function can mess the counters. See how we call `self.counter_check()` after `assert_status()` in `test_scm_status_cache`.

# This Diff

By returning the total number of tries from `assert_status()`, we can easily check the counters based on it because we now know how many misses happened ( `num_of_tries - 1`).

Reviewed By: kmancini

Differential Revision: D61643516

fbshipit-source-id: b973a2f006e93f44cf8fd7fb3f322024f8cd4d4b