mononoke_macros: move just_knobs.json to mononoke_macros/just_knobs_defaults
Summary: `just_knobs.json` is a file that we currently use to override just knobs in integration tests. I want to use this file to override just knobs in unit tests too which requires it to live in the `mononoke_macros` directory. This diff moves it to `mononoke_macros/just_knobs_default` and adds it to the `mononoke_macros` buck target, and modifies integration tests setup logic accordingly

Reviewed By: lmvasquezg

Differential Revision: D61659286

fbshipit-source-id: be7f7ad2f46cc3fbeef3764d83d486672ce1b9dd