Unit tests demonstrating the failure in write text and data channel
Summary:
When writing packetline data in git protocol, the maximum amount of data that can be written per packetline line is 65516 bytes. Due to this restriction, we split the incoming input into chunks of maximum size 65516 bytes before writing to output. However, when writing the data as text or when writing to a particular band, we prefix every line with a one byte character (`\n` in case of text and the band byte in case of data channel). This character puts the size over the threshold leading to write failure.

This diff demonstrates this problem through unit tests. I will fix this in the next diff and update the tests accordingly.

Reviewed By: mzr

Differential Revision: D61712236

fbshipit-source-id: 85f7b1ab8567495ace1005574a089c261e01e77f