Unit test demonstrating the failure in write text and binary reporting
Summary: The number of bytes written reported by `write_text_packetline` and `write_binary_packetline` are incorrect. While `write_text_packetline` over-reports the count due to the addition of newline characters, `write_binary_packetline` under-reports the count due to the lack of prefix while writing the packetline. This diff demonstrates that problem which will be fixed in the next diff.

Reviewed By: mzr

Differential Revision: D61713741

fbshipit-source-id: 4bf5f9fdd06563f9e77b976ec06a80dfd3275e9a