scuba_ext: add FutureStatsScubaExt::log_future_stats
Summary:
Timing a future and then immediately logging its results to scuba is a common operation, so much so that we've grown a couple of utility functions that try to make that more ergonomic.

Create an extention trait that lets us log the result of `TimedFutureExt::timed` and `TryTimedFutureExt::try_timed` directly to scuba with a single method call.

Reviewed By: lmvasquezg

Differential Revision: D61715430

fbshipit-source-id: b0cf0f8d1486683c9d6299131fec5627ef97eead