mononoke_types: introduce types for SkeletonManifestV2
Summary:
`SkeletonManifestV2` is a sharded manifest type intended to replace `SkeletonManifest`. `SkeletonManifestV2` differs from `SkeletonManifest` in the following ways:
1) It doesn't include any extra metadata except a rollup count of its descendants  that's used to enable ordered traversal. In particular this means that it can no longer be used for case conflicts checking which will be delegated to a different derived data type (`CaseConflictsManifest`).
2) It's using the same sharding structure as BSSM, which can be conceptually thought of as the entire tree being stored in a single sharded map, were the sharded map decides how the tree is split up into blobs with no respect for the directory structure. This avoids having many tiny blobs corresponding to every small directory which can hinder read and write performance.

This makes `SkeletonManifestV2` identitical to `BssmV3` with the exception of not doing the transform that prepends the basename suffix to the paths.

I'm using a smaller sharded map weight limit (500 instead of 2000) compared to BSSM. The larger the weight limit the fewer the number of blobs which can result in faster traversal for large portions of the tree and faster derivation for commits that touch a large number of files. On the other hand the larger the weight limit the larger the blobs which can make derivation and lookups on the average case slower, and will increase the total storage required by the manifests. This can be adjusted as appropriate in early experimentation.

Differential Revision: D61712639

fbshipit-source-id: cbdb601fdad9a6830268a797f26cd0ff4923ad4b