Extend HookManager to check for Git repo hook bypasses
Summary:
For Sapling repos, bypassing all hooks can only be done by members of SCM group. However, for git repos there is a dedicated action on every repo ACL called `write_no_hooks`. Members of the ACL having access to this action should have the ability to bypass all the hooks for a given repo. To maintain feature parity with Metagit, this diff provides that functionality.

The next diff in the stack will hook it up with Mononoke Git server so we will be able to test it E2E. In this diff, we are relying on existing integration tests to detect potential regression.

Reviewed By: mzr

Differential Revision: D61723920

fbshipit-source-id: 86ff06e6feff93ef6331a5afcb362171b9cf6bbb