Add HorizontallyGrowingTextField and use for quick commit title
Summary:
Add a new component that looks like `<TextField>` but grows horizontally as you type. This is often useful so the field isn't super long in the UI unless you're actively typing something long there. We also can have a max width so it doesn't get too crazy.

The implementation of this is a little tricky. We have a container `<div>`, with the `<input>` inside. The input sets data-value of the parent div on each keystroke in javascript. The parent div has an `::after` with `content: attr(data-value)` and `visibility: hidden`, which means the ::after gets an invisible copy of the text content. This ::after then resizes the container div, which means the <input> gets just enough space to fit all the text.

This approach is nice because the javascript is very simple, and we don't need to do any expensive DOM size measurements (an alternate approach would be to have a sibling `<span>` that is given the text content in js, then we copy the width of the span to the input. That's more expensive to do each keystroke).

Reviewed By: quark-zju

Differential Revision: D61816960

fbshipit-source-id: 07c9975c7fd3d7ec702238b92a75ba2088122d08