tests: don't set default value for remotefilelog.reponame
Summary:
We have logic to infer reponame based on paths.default, so in general it does not need to be specified explicitly.

It is important for the reponame to be correct in order for SLAPI to work w/ Mononoke. The Mononoke integration tests do a lot of futzing with reponame, and it will be a lot simpler and correct to let Sapling infer it.

Reviewed By: lmvasquezg

Differential Revision: D61679639

fbshipit-source-id: 8f71cfca23dda3cbd7071b6f713da44ee8beee49