tests: kill hgclone_treemanifest
Summary:
Replace with a plain "hg clone".

The main reason I'm doing this is to set the remotefilelog.reponame properly based on the name of the repo being cloned. Previously it was set to the name of the destination directory, which was incorrect and resulted in a non-functional SLAPI.

If you look carefully you will see I'm not actually setting remotefilelog.reponame to anything. This is because Sapling by default will infer remotefilelog.reponame based on paths.default.

Reviewed By: lmvasquezg

Differential Revision: D61643914

fbshipit-source-id: a69c781155f7d38d6f1ea5a5a29cc8722b3e0a20