Fix submodule deletion on merge commits
Summary:
Fixes issue from the unit test added in previous diff (D61840712).

The problem is that, when deleting a submodule, we're currently assuming that all the parents have this submodule in their working copy and we fail if that's not the case.

But in reality, we just need one of the parents to have it.

Reviewed By: andreacampi

Differential Revision: D61839040

fbshipit-source-id: 438b7b45c6c547ccafd1bcf83ca6c0f9386c29f7