derived_data_tailer: fix hot config reloading
Summary: Hot config reloading works on the level of the `MononokeRepos` object. Storing a repo object or a repo facet will mean that hot config reloading will not affect it. This diff restructures the derived data tailer so that in each iteration of the tailer we fetch the repo object from `MononokeRepos`, which should be the repo object built using the most recent config.

Reviewed By: RajivTS

Differential Revision: D61848324

fbshipit-source-id: d6caddc8eb1fb55e76c9697d522ff15d19e91d6b