extract test log command to bash helper
Summary:
# This stack
Setting up a robust integration test to test Mononoke Git server with xrepo infra and submodule expansion.
This will test switching the source of truth from Mononoke Git to large repo.

## This diff
As title.
I'm also setting the `REPONAME` environment variable to large repo, to avoid errors when we run `sl` commands and it fails because it can't find the changeset (because it looks for it in the wrong repo id).

Reviewed By: andreacampi

Differential Revision: D61597476

fbshipit-source-id: f93ce40bcee8bf659a5b471047e47dd43c15fbb3