Extract hook integration test setup into separate function
Summary:
# This stack
Setting up a robust integration test to test Mononoke Git server with xrepo infra and submodule expansion.
This will test switching the source of truth from Mononoke Git to large repo.

## This diff
Extracting the logic to register a hook to a separate function, so it can be used in future diffs.

Reviewed By: andreacampi

Differential Revision: D61597477

fbshipit-source-id: 462bae2c6c1abcb1de8eb5b91968cedf862dc8e7