Create large repo commit in xrepo submodule integration test setup
Summary:
# This stack
Setting up a robust integration test to test Mononoke Git server with xrepo infra and submodule expansion.
This will test switching the source of truth from Mononoke Git to large repo.

## This diff
Because it's good to have at least one commit in the large repo before syncing/merging, I started adding one on top of some integration tests.

But that broke consistency with other tests.
I'm adding another one and noticed that I would copy paste this in the test again, so it's best to just do it in the common set up for all tests.

I added it to the common setup and removed the copy-pasting from the existing tests where I added the commit.

In the process, I removed some hardcoded values from existing tests as well, so they get bonsai ids from the output.

Reviewed By: andreacampi

Differential Revision: D61478630

fbshipit-source-id: c16aa49cddaa00f6249fb09e2beaf7ea9c8f0780