Helper to wait for bookmark moves in mononoke git repos
Summary:
# This stack
Setting up a robust integration test to test Mononoke Git server with xrepo infra and submodule expansion.
This will test switching the source of truth from Mononoke Git to large repo.

## This diff
Added a helper function to wait for bookmark moves in a Mononoke repo using bonsai changeset ids, instead of mercurial ones.
This is needed because in further diffs we'll need to wait for bookmark moves in small repos imported from git, which don't derive hg types because they have submodules.

I also parameterized the max number of attempts in the existing `wait_for_bookmark_move_away_edenapi`. I didn't use it in the end, but figured it could be useful in other tests.

Reviewed By: andreacampi

Differential Revision: D61597861

fbshipit-source-id: 3f81d99f88d9efc1cf6dd6c9a98568407eb04f7f