fix a bug when sub command name can be parsed wrong
Summary:
Found this when doing experiments locally. I have my dev alias `eden-dev` which will carry `--config-dir <dir-path>` every time it gets called. So the code will think "<dir-path>" being the name of the sub command if we simply skip the first argument and look for the first string not starting with "-" as the original code below
```rust
    let subcommand_name = args
        .skip(1)
        .find(|a| !a.starts_with('-'))     # <---- fail for the case when carrying e.g. "--config-dir some_dir_path"
        .ok_or(anyhow!("missing subcommand"))?;
```

Changing it to a regex so it's more accurate when parsing.

Note:
It's currently working fine even if you explicitly append "--config-dir <dir-path>" in your `eden` command because we have [empty list](https://fburl.com/code/mj2h42ov) of `EXPERIMENTAL_COMMANDS` so `is_command_enabled_in_rust` will always return true for a command already defined in Rust.
Without this change, if we define `remove` command in Rust and add it to `EXPERIMENTAL_COMMANDS`, `is_command_enabled_in_rust` will return True (due to the wrong subcommand name parsed) when it should return False.

Reviewed By: kmancini

Differential Revision: D61005404

fbshipit-source-id: 308e6bcc48d9981bb5fa728c5406929632625116