fix status cache counter check for windows
Summary:
## Background

We added the diff to apply an optimization to skip .hg dir changes when checking if it's ok to reuse the cached status resutls. Some tests start failing under mode opt-windows. After triaging we found that the failure was caused by the delayed sequence number increase. The sequence number was increased during our counter check which was unexpected and broke the assumption of the original tests. The interesting part is that this seems to be the issue only for this specific mode.

## This Diff

To improve the stability of our tests, this diff adds an explicit sync before start requesting for status.

Reviewed By: genevievehelsel

Differential Revision: D61863288

fbshipit-source-id: 92673556e26470f6ceb6aceba60ed159d6841ed8