automv: increase the value of automv.max-files
Summary:
We introduced the config in D30052694 due to the poor performance of auto move
detection. However, the default value (50) is too small and can miss moves when
there are files more than the threshold. I tested a commit with 1000 files, and
it can finish within a few seconds. So this diff sets the default to 1000.

In the long term, we should fix the poor performance of the auto move detection
algorithm. Currently, the complexity of the algorithm is O(n^2), we can
optimize it to O(n). We already have implemented similar algorithm in copy
tracing project. We need to refactor the code and expose it to Python, so we
can use it in this auto move detection.

Reviewed By: sggutier

Differential Revision: D61752434

fbshipit-source-id: 5dc8c19a616602f875325df0359f8e144bcbb2fb