commit cloud: retrieve timestamps as secs
Summary:
**Short explanation:** Since now we send seconds to the client since D61660595 let's also write and read seconds in our tables.

**Long explanation:** Basically D61660595 changed things so when we return to the client a timestamp we return the seconds part of a timestamp (that it's stored with nanoseconds precision) e.g  if you have 1724801478437195146 we return 1724801478 so the client can handle the timestamp. And it works fine in tests because in sqlite we store nanoseconds and read nanoseconds.

The thing I missed was that in prod, we __write nanoseconds__ but we __read seconds__, this cause we use UNIX_TIMESTAMP which downgrades the timestamp to seconds, so we were creating a Timestamp with nanoseconds precision but when retrieved, only had seconds, and then we were trying to get seconds out of that. So let's say we stored 1724801478437195146, we read 1724801478 and then we returned to the client the "seconds part of it"... meaning like 1.

So good news is that we have the right info, the bad news is that when you do `hg cloud sl --history` it says workspaces were saved in 1970... not cool.

The solution is handling timestamps as seconds both in reads and writes

Reviewed By: zertosh

Differential Revision: D61859958

fbshipit-source-id: 9868c71ae3b24b9c6180e64a34a58b6f315bdcf6