derived_data_service: restructure derivation workers to fix hot config reloading
Summary:
Hot config reloading currently doesn't work for derivation workers because it operates on the level of the `MononokeRepos` object. Storing a repo object or a repo facet will mean that hot config reloading will not affect it. Changing the derivation worker code to rely on `MononokeRepos` and fetching a new repo object in each iteration fixes this.

But just relying on `MononokeRepos` isn't enough in this case. In the derivation worker code we spawn a `Coordinator` that corresponds to a single repo and config name. This doesn't work with hot config reloading as we need another layer to control spawning/stopping new tasks depending on added or deleted derived data configs. This diff renames the current `Coordinator` struct to `Worker`, and adds another `Coordinator` struct that handles all the different configs for a repo and spawns a `Worker` for each config, adjusting the workers as derived data configs are added/deleted.

Reviewed By: andreacampi

Differential Revision: D61858001

fbshipit-source-id: 28e63da37b4b64072174820f86dfa64aea12bf9a