Fix insidious push-poller issue that caused it to always hit Metagit
Summary:
For every repo that has SoT in Mononoke, the push-poller is responsible for  checking the fingerprint of the repo in Mononoke and updating the Metagit DB so that Metagit followers can start replicating from Mononoke.

While performing a SoT switch today, we realized that the push poller was comparing Metagit's fingerprint against Metagit itself instead of Mononoke. This was because the extra git header that we use to redirect traffic to Mononoke was not being honored by the command. The reason for this was that we were specifying single quotes for the argument which are not supported in the args. We needed to provide the header argument without it which is what this diff does.

Reviewed By: singhsrb

Differential Revision: D61909468

fbshipit-source-id: 65456b2e985890847d44f0cb2801f845e058cd05