tests: use same repo name for hg and mnke
Summary:
Using the same name should allow the removal of "hgmn" and in turn "runtime" REPONAME usage since the "hg" repo won't have to change its repo name when it wants to talk to the mononoke repo.

It might be a bit confusing having multiple things called "repo", but in general I think it is okay to have the same name since they do truly represent the same repo. I think it will be simpler overall than the REPONAME and hgmn stuff all over the place.

Reviewed By: gustavoavena

Differential Revision: D61703783

fbshipit-source-id: e8ad1a899106d4c4ed0df853ff432da81cc5697b