Log the queue we are using to serve the request
Summary:
I am struggling to understand the data I'm seeing when cross-referencing JK, Scuba and ODS. Some logging will help.

This is noisy, but it is conditional on `--thift-server-mode=factory-group` which is currently set only on the test tier. I will remove the logging later when we are ready to move to prod.

Reviewed By: gustavoavena

Differential Revision: D61915516

fbshipit-source-id: e053ce7fafa588403a0c593f51de1e37439285ce