Fix for Mononoke Git integration test for push
Summary: This particular integration test was flaky because it depended on WBC for catching up with the latest changes. Usually that happens, but its a timing problem and sometimes it might lag which would cause the test to fail making it flaky. The changes in this diff remove the timing dependency so now it should always work

Reviewed By: lmvasquezg

Differential Revision: D61967592

fbshipit-source-id: 6aa547042a76d5da34adc685a7edbc9e8a56029f