commit cloud: always fetch local and remote bookmarks
Summary:
**Expected Behavior (old commit cloud service)** : When forming a smartlog, we check each draft commit to determine if there is a local or remote bookmark associated with it, and then add that information to the node. Additionally, if  `ADD_ALL_BOOKMARKS` is provided, we include nodes in the smartlog that have a remote/local bookmark associated with them, even if they are not draft or user commits.

**Current behaviour (bug)**: if `ADD_ALL_BOOKMARKS` was provided, we fetched bookmarks and added them to the smartlog.

However, if the flags __were not provided__, we did not add **any** bookmarks to the smartlog at all. This approach was problematic because, without such flag, users would not see any bookmarks associated with a commit, which was inconsistent with the functionality of the old commit cloud service.

I have corrected this so that the logic is now accurately reflected.

Reviewed By: andreacampi

Differential Revision: D61934443

fbshipit-source-id: a8a69782540d8161905cceedf8ea2051a214bf2d