Group all git-related RepoConfigs
Summary:
## This stack
I'll be adding new fields to the repo configs that are related to Git servers and cross-repo sync.
Before doing that, I thought it would be better to group all of the git configs together, to avoid having them randomly spread across the `RepoConfig`.

The plan is
1. Add the new `GitConfigs` (and RawGitConfigs) structs and set them in all configs. **Don't stop using the old ones yet**
2. Sync to fbsource and start reading from the new field.
3. Deprecate the old fields (once the fbsource changes have been pushed to all tiers).

## This diff
This just introduces the `GitConfigs` struct in fbsource layer and starts using it after parsing the config.
It's backwards compatible, since it's still reading from the old fields.

After D61969781 has been pushed to configerator, we'll start **reading from the new `git_configs` field** (D61966808, after thriftr changes from D61969781 are synced to fbsource)

Reviewed By: YousefSalama

Differential Revision: D61966506

fbshipit-source-id: 112c4a4100469b86b54c3ff90ddfc89a78862e9b