tests: use "mono:" scheme everywhere
Summary:
My goal is to make clones from an hg server automatically switch to mononoke once mononoke is started.

This is achieved by first pointing the "mono:" scheme to "ssh://user@dummy/REPO", and then when mononoke starts, it overrides the "mono:" scheme to instead point to "mononoke://$(monononke_address)/REPO". This only works if the REPO name is identical (which it is now, after my previous change).

Reviewed By: gustavoavena, andreacampi

Differential Revision: D61703781

fbshipit-source-id: 59aa744068f0460b5f6faba7acada2f74394f629