tests: kill hgmn() function
Summary:
This was an adapter to point the current repo at Mononoke (instead of, presumably, an hg server repo).

After recent changes, this should no longer be necessary. We use a "mono:" scheme as an indirection point for the repo source, which can be changed automatically after mononoke starts up.

The assumption here is that once mononoke is started, you never really want to talk to the hg server repo anymore.

Reviewed By: andreacampi

Differential Revision: D61703779

fbshipit-source-id: 7d0c0bea2fb3506962e019938b288f8ddac3dbbe