backout: fix backout of copied file
Summary:
The copy handling code was assuming both sides of a copy would show up in the manifest diff. This isn't the case, though, if the source of the copied file was not modified.

Fix by checking whether the other side of the copy is in the diff or not.

This issue appeared after D60698882 stopped filtering out the copy in this case.

Reviewed By: zzl0

Differential Revision: D61944971

fbshipit-source-id: 03f82e8e37327c70a5f3542c30762bcb97e03766