git_server_push_poller: fix deadlock with large data
Summary:
We noticed that using the poller for `yocto/meta-openbmc` results in a deadlock
in the poller.

Turns out this was due to poller dealing with a large amount of data since this
repository has around 8k refs. This situation leads to overflowing the I/O
buffers used while working with stdin and stdout streams before this diff
because the underlying OS then refuses to process further data until there is
some processing of the data already existing on the streams. However, since we
were doing all the processing on streams on the same thread, we end being
deadlocked in the situation because we are effectively waiting for data to
completely written before our thread will start reading it.

The solution is to ensure that all the streams are unblocked to work with data
to ensure the I/O buffers are constant being written or read to to avoid the
buffer filling up situation.

Reviewed By: RajivTS

Differential Revision: D62016461

fbshipit-source-id: 16d651cfd4ebaf569e96c0e91d93af3eb070f405