Make the key prefix check forward- and backward-compatible
Summary:
## This stack

Make `async_requests` general-purpose so that we can schedule more async work. The initial consumer will be SCS, but I expect this might be useful even more broadly than that.

## This diff

The prefix check only recognizes the current key schema. That is unnecessairly strict, because in many cases we can still handle other entries.

This diff adds an allowlist of other key prefixes we can handle, and adds a key that was added in D61848170 and backend out in D62020463, leaving table entries that we can't process otherwise.

Reviewed By: gustavoavena

Differential Revision: D62024563

fbshipit-source-id: 50da25a42e366c7daa398049ba64f8a7fe956763