Create BookmarkHook, distinct from FileHook and ChangesetHook
Summary:
`FileHook` runs on all files in a commit.
`ChangesetHook` runs once per commit.

Both have the commonality that they are executed on each new commits introduced by a bookmark move.

As we support Git hooks, we now need to support a separate kind of hook that acts on the bookmark move itself.
It must run on bookmark creation or move even if the bookmark move doesn't introduce any new changeset.

The three examples of such hooks that already exist are:
* `block_new_bookmark_creations_by_name`
* `block_accidental_new_bookmark_creation`
* `block_unannotated_tags`
which this diff migrates to the new mechanism.

This will avoid the unfortunate situation where previously, a bookmark move that doesn't introduce new changesets would be skipped.

Reviewed By: RajivTS

Differential Revision: D61984306

fbshipit-source-id: cafe6211776352adb1bb634a9aff1af25c88f40c