Expose problems with hooks that apply to bookmarks
Summary:
Use the same jk that's currently used in prod for all repos, which will become the default without a jk.

Show that the behaviour of hooks that apply to bookmark is unexpected:
* Running them can be omitted in certain situations, as shown in `test-mononoke-git-server-block-unannotated-tags.t`
* They would require all commits pushed with this bookmark to have the annotation to bypass them, as shown in both other modified tests

These issues get fixed in the child diff by handling hooks that apply to bookmarks differently from hooks that apply to changesets or files, so that each category of hook can be run appropriately.

Reviewed By: YousefSalama

Differential Revision: D61984305

fbshipit-source-id: 130f7d5d117629b251f2686808f83c22e69c82ef