Send binary payload in base64 in a single message
Summary:
Previously, image uploads were accomplished by sending two messages from client to server:
1. normal text message in JSON providing the filename and signaling that the next message will be binary
2. a binary message containing only the file content

We've seen reports of people only getting 1/2 of these messages delivered, causing stuck uploads.

If we can send everything in one message, it will be simpler and more robust.

We could either stuff the image content in base64 into the JSON message, or stuff the JSON message into the binary message.

I think for simplicity and consistency, it makes a lot more sense to use base64 encoding of the image data.

This diff removes the binary message handling for image uploads, and converts the image file to base64 to send in the normal message.

You would typically expect there to be a 33% increase in message size due to this. Sending the message from the client to the server is part of the latency of uploading an image. However, I do not believe this is the main slowdown, so I don't think this diff leads to 33% slower image uploads.

Our analytics are measuring the server upload time, not including this transport time, but I guess we could change that if we really need to profile this.

In the next diff, I'll remove vestigial remains of binary messages, so everything is only strings and we don't need special code paths.

The image upload tests also became simpler/more clear after doing this

Reviewed By: muirdm

Differential Revision: D62002379

fbshipit-source-id: fc9616ce380412b4caaa22300ed8ca488dfb173b