Allow scrolling tooltip dropdowns that would go off screen
Summary:
As we've added more settings over time, it's become possible for the setting dropdown to be too tall to fit on screen.

We need to update Tooltip to set a maxHeight computed when it's created based on the screen height and the tooltip content height. If it's too tall, set `height: 100%` so the tooltip contents may decide to scroll themselves.

Then, the DropdownFields component used by most of the TopBar buttons can use overflow: scroll to get nice scrolling.

Reviewed By: muirdm

Differential Revision: D61998023

fbshipit-source-id: 42936e8b37d36471227035bf9b8be2f846199baf