Determine if a commit is irrelevant to your cwd
Summary:
Now that we have computed the common path prefix for a commit, we can use it to determine if that commit is irrelevant to your cwd.

This is useful for large monorepos. If your cwd is `/addons/isl`, then commits containing files entirely within `/edenfs/` are irrelevant to you. If you change your cwd to `/addons`, then commits with changes inside `/addons/isl` are relevant and edenfs is not.

The specific logic of this is relatively simple, but also fairly specific. The code basically requires either `cwd.startsWith(prefix)` or `prefix.startsWith(cwd)` in order to be relevant.

For example, if a commit changed files like
```
addons/isl/README.md
addons/isl/src/index.ts
```
the common prefix would be `addons/isl` for that commit. Then:

**Repo root as cwd means everything is relevant**:
If your cwd was the repo root, aka `''` (when expressed as a repo-relative path), then since `addons/isl` starts with `''`, this commit is relevant. This means if you do not use a cwd other than the repo root you will never see this feature.

**matching prefix and cwd is relevant**
If your cwd was `addons/isl`, then again this commit is relevant because cwd === prefix.

**cwd "above" changes is relevant**
If your cwd was `addons`, then prefix startsWith cwd, so it is relevant

**cwd "below" changes is relevant**
If you cwd was `addons/isl/src/foo`, then cwd startsWith prefix, so it is relevant. This is because changes "above" you may be relevant to you. This is not really guaranteed and is a source of false-positives. It's better to falsely say something is relevant to you than to falsey say it's not, IMO.

**cwd with a difference to prefix is irrelevant**
The only case then for irrelevant commits is if there is a material difference between the paths.
In this example, if cwd was `addons/vscode`, then the `addons/isl` commit would be irrelevant.

One important property of this system is that if you have a commit with only irrelevant changes, but add a single relevant change, then it should become relevant. Since the commonality of the irrelevant and one new relevant file will by definition be a subset of the relevant file's path, then the new prefix will match the cwd.

Reviewed By: muirdm

Differential Revision: D61988978

fbshipit-source-id: 5df3b4812802530073eb4e48164164a6f5cd0cfe