Deemphasize irrelevant commits
Summary:
Now that we know if a commit is irrelevant, add UI to render it that way.

- grey out irrelevant commits in the dag
- add a special icon to represent this state. We will use this icon elsewhere for consistency. It's an IconStack of a folder with an 'x' in the corner.
- if you hover on the icon, you get a clear description about why this commit is considered irrelevant.

The design here is intended to mean this commit has something about its folder that makes it different, which is shown via the icon. When combiend with greying out, it reads as "irrelevant due to the directory" to me.

My only hesistation is that the folder icon is on the left, and could maybe be missed. It seems like it stands out to me, but if you miss that icon, you might think it's obsolete (in sl terms) instead of irrelevant (in cwd terms)

In a follow up diff, we'll add a config setting to disable this feature for folks who don't like it.

Differential Revision: D61988979

fbshipit-source-id: f0cd730725b227a982e877a5271476be0b747dc3