Add maxCommonPathPrefix to CommitInfo
Summary:
Add a new field to CommitInfoView that tracks the longest common path prefix of all the files in the commit. This must be computed on the server before truncating the file list, so that it accounts for all files in the commit.

This will be useful for determining on the client if a commit is relevant to your cwd. The use case at hand is that you mount a cwd like `/www/`, we should be able to deemphasize any commits that only modify files in `/fbcode/` as being irrelevant. This is helpful for monorepos where you may want to focus on a specific subfolder at a given time. This deemphasizing is not implemented in this diff.

Why compute it this way? If we instead tried the more direct approach of computing `isCommitRelevantToCwd`, we'd need to use the cwd in the computation of the CommitInfo. However, we share fetched CommitInfos among multiple clients of the server, meaning it would not be accurate for all clients. Instead, it's trivial to compute isCommitRelevantToCwd given the cwd and maxCommonPathPrefix.

Reviewed By: muirdm

Differential Revision: D61743900

fbshipit-source-id: 2a30ed109e58e7f5fab31c759adb32929c601178